requirements.txt
torch>=1.12.0
numpy>=1.21.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
colour-science>=0.4.3
scipy>=1.11.0
rawpy>=0.18.0
imageio>=2.28.0
OpenEXR>=1.3.9
# Deep learning colorization dependencies
Pillow>=9.0.0
torchvision>=0.13.0
huggingface_hub>=0.15.0
timm>=0.9.2

